// web/src/lib/api/generated/models/meetingScopeRequest.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { MeetingScopeRequestDeletion } from "./meetingScopeRequestDeletion";

export type MeetingScopeRequest = unknown &
  unknown &
  unknown & {
    after?: string;
    before?: string;
    deletion?: MeetingScopeRequestDeletion;
    /**
     * @minItems 1
     * @maxItems 100
     */
    domains?: string[];
    /**
     * @maxItems 100
     * @items.minimum 1
     * @items.maximum 9007199254740991
     */
    message_ids?: number[];
    /**
     * @minimum 1
     * @maximum 9007199254740991
     */
    participant_id?: number;
    /**
     * @minItems 1
     * @maxItems 100
     * @items.minimum 1
     * @items.maximum 9007199254740991
     */
    participant_ids?: number[];
    /**
     * @minimum 1
     * @maximum 9007199254740991
     */
    person_id?: number;
    /**
     * @minItems 1
     * @maxItems 100
     * @items.minimum 1
     * @items.maximum 9007199254740991
     */
    source_ids?: number[];
  };